feat(client): :sparkles: make sponsor clickable
