+++ client/src/components/Pages/dispositif/Edition/SponsorsEdit/SponsorsEdit.module.scss
@@ -12,4 +12,5 @@
 .add {
   margin-left: u(10);
   box-shadow: none !important;
+  text-wrap: nowrap;
 }
