+++ client/src/components/Pages/dispositif/Sponsors/Sponsors.module.scss
@@ -38,6 +38,9 @@
   display: flex;
   align-items: center;
   max-width: 200px;
+  padding: u(1) u(2);
+  border-radius: u(2);
+  cursor: pointer;
 
   @include media-max("lg-limit") {
     @include dsfr-text("normal");
@@ -47,13 +50,12 @@
   }
 
   &.edit {
-    padding: u(1) u(2);
-    border-radius: u(2);
     border: 1px solid $lightBorderDefaultGrey;
-    cursor: pointer;
-    &:hover {
-      background: $lightBackgroundContrastGrey;
-    }
+    margin-bottom: u(1);
+  }
+
+  &:hover {
+    background: $lightBackgroundContrastGrey !important;
   }
 }
