+++ client/src/components/UI/RichTextInput/lib.ts
@@ -24,19 +24,6 @@ export function getSelectedNode(selection: RangeSelection): TextNode | ElementNo
 
 }
 
-export function sanitizeUrl(url: string): string {
-  const URL_PROTOCOL = /^(?:https?|mailto|ftp|tel|file|sms)/gi;
-
-  /** A pattern that matches safe  URLs. */
-  const SAFE_URL_PATTERN = /^(?:(?:https?|mailto|ftp|tel|file|sms):|[^&:/?#]*(?:[/?#]|$))/gi;
-
-  url = String(url).trim();
-
-  if (!url.match(URL_PROTOCOL)) return `https://${url}`;
-  if (url.match(SAFE_URL_PATTERN)) return url;
-
-  return "https://";
-}
 
 // Source: https://stackoverflow.com/a/8234912/2013580
 const urlRegExp = new RegExp(
